requirements.txt
asgiref==3.2.3
click==7.1.1
Django==3.0.4
django-contact-form==1.8.1
django-crispy-forms==1.9.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-simple-menu==1.2.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
Werkzeug==1.0.0
